APEXCORE-617 InputNodeTest intermittently fails with ConcurrentModificationException
